// frontend/generated/openapi-client/types.gen.ts
export type CourseReview = {
    id: string;
    content: string;
    rating: number;
    userId: string;
    courseId: string;
    instructorReply?: string;
    updatedAt: string;
    createdAt: string;
    user: User;
    course: Course;
};

export type CourseComment = {
    id: string;
    content: string;
    userId: string;
    questionId: string;
    updatedAt: string;
    createdAt: string;
    user: User;
    question: CourseQuestion;
};

export type CourseQuestion = {
    id: string;
    title: string;
    content: string;
    userId: string;
    courseId: string;
    updatedAt: string;
    createdAt: string;
    comments: Array<CourseComment>;
    user: User;
    course: Course;
};

export type CourseCategory = {
    id: string;
    name: string;
    slug: string;
    description?: string;
    updatedAt: string;
    createdAt: string;
    course: Array<Course>;
};

export type CourseEnrollment = {
    id: string;
    userId: string;
    courseId: string;
    enrolledAt: string;
    updatedAt: string;
    createdAt: string;
    user: User;
    course: Course;
};

export type Course = {
    id: string;
    slug: string;
    title: string;
    price: number;
    level: string;
    status: string;
    description?: string;
    isPublished: boolean;
    instructorId: string;
    thumbnailUrl?: string;
    discountPrice?: number;
    shortDescription?: string;
    updatedAt: string;
    createdAt: string;
    sections: Array<Section>;
    lectures: Array<Lecture>;
    reviews: Array<CourseReview>;
    questions: Array<CourseQuestion>;
    categories: Array<CourseCategory>;
    enrollments: Array<CourseEnrollment>;
    instructor: User;
};

export type Account = {
    id: string;
    userId: string;
    type: string;
    provider: string;
    providerAccountId: string;
    refresh_token?: string;
    access_token?: string;
    expires_at?: number;
    token_type?: string;
    scope?: string;
    id_token?: string;
    session_state?: string;
    user: User;
};

export type Session = {
    id: string;
    sessionToken: string;
    userId: string;
    expires: string;
    user: User;
};

export type User = {
    id: string;
    bio?: string;
    name?: string;
    email?: string;
    image?: string;
    emailVerified?: string;
    hashedPassword?: string;
    courses: Array<Course>;
    accounts: Array<Account>;
    sessions: Array<Session>;
    courseReviews: Array<CourseReview>;
    courseComments: Array<CourseComment>;
    courseQuestions: Array<CourseQuestion>;
    lectureActivities: Array<LectureActivity>;
    courseEnrollments: Array<CourseEnrollment>;
};

export type LectureActivity = {
    id: string;
    userId: string;
    lectureId: string;
    progress: number;
    isCompleted: boolean;
    lastWatchedAt: string;
    user: User;
    lecture: Lecture;
};

export type Lecture = {
    id: string;
    order: number;
    title: string;
    duration?: number;
    description?: string;
    courseId: string;
    sectionId: string;
    isPreview: boolean;
    videoStorageInfo?: {
        [key: string]: unknown;
    };
    updatedAt: string;
    createdAt: string;
    activities: Array<LectureActivity>;
    course: Course;
    section: Section;
};

export type Section = {
    id: string;
    order: number;
    title: string;
    description?: string;
    courseId: string;
    updatedAt: string;
    createdAt: string;
    lectures: Array<Lecture>;
    course: Course;
};